Refactor configs into DB layer
